api-smatic-dev/requirements.txt
Django==4.0.5
django-ninja==0.18.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pyjwt==2.4.0
pytz==2022.1
paho-mqtt==1.6.1